Refactor Item to TS and split out CartStatus
